Refactor View License Summary services and presenters
Extract the common view license data (data not in the View License tabs) into it's owm service and presenter.

Use the new view license and view license summary functionality called from the View License Summary Controller.

Optimise the fetch calls to get licence data per service requirement.
