+++ app/controllers/licences.controller.js
@@ -7,6 +7,7 @@
 
 const InitiateReturnRequirementSessionService = require('../services/return-requirements/initiate-return-requirement-session.service.js')
 const ViewLicenceService = require('../services/licences/view-licence.service.js')
+const ViewLicenceSummaryService = require('../services/licences/view-license-summary.service')
 
 async function noReturnsRequired (request, h) {
   const { id } = request.params
@@ -28,11 +29,13 @@ async function viewSummary (request, h) {
   const { params: { id }, auth } = request
 
   const data = await ViewLicenceService.go(id, auth)
+  const summaryData = await ViewLicenceSummaryService.go(id)
 
   return h.view('licences/view.njk', {
     activeNavBar: 'search',
     activeTab: 'summary',
-    ...data
+    ...data,
+    ...summaryData
   })
 }
