+++ test/controllers/licences.controller.test.js
@@ -14,6 +14,7 @@ const Boom = require('@hapi/boom')
 // Things we need to stub
 const InitiateReturnRequirementSessionService = require('../../app/services/return-requirements/initiate-return-requirement-session.service.js')
 const ViewLicenceService = require('../../app/services/licences/view-licence.service.js')
+const ViewLicenceSummaryService = require('../../app/services/licences/view-license-summary.service')
 
 // For running our service
 const { init } = require('../../app/server.js')
@@ -164,7 +165,8 @@ describe('Licences controller', () => {
 
     describe('when a request is valid', () => {
       beforeEach(async () => {
-        Sinon.stub(ViewLicenceService, 'go').resolves(_viewLicenceSummary())
+        Sinon.stub(ViewLicenceService, 'go').resolves(_viewLicenceData())
+        Sinon.stub(ViewLicenceSummaryService, 'go').resolves(_viewLicenceSummary())
       })
 
       it('returns the page successfully', async () => {
@@ -177,10 +179,14 @@ describe('Licences controller', () => {
       })
     })
 
+    function _viewLicenceData () {
+      return {
+        licenceRef: '01/130/R01'
+      }
+    }
     function _viewLicenceSummary () {
       return {
         id: '7861814c-ca19-43f2-be11-3c612f0d744b',
-        licenceRef: '01/130/R01',
         region: 'Southern',
         startDate: '1 November 2022',
         endDate: '1 November 2032'
