+++ test/presenters/licences/view-licence-summary.presenter.test.js
@@ -8,22 +8,20 @@ const { describe, it, beforeEach } = exports.lab = Lab.script()
 const { expect } = Code
 
 // Thing under test
-const ViewLicencePresenter = require('../../../app/presenters/licences/view-licence.presenter.js')
+const ViewLicenceSummaryPresenter = require('../../../app/presenters/licences/view-license-summary.presenter')
 
-describe('View Licence presenter', () => {
+describe('View Licence Summary presenter', () => {
   let licenceAbstractionConditions
   let licence
-  let auth
 
   beforeEach(() => {
     licence = _licence()
     licenceAbstractionConditions = _abstractionConditions()
-    auth = undefined
   })
 
   describe('when provided with a populated licence', () => {
     it('correctly presents the data', () => {
-      const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+      const result = ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
       expect(result).to.equal({
         abstractionConditionDetails: {
@@ -40,26 +38,19 @@ describe('View Licence presenter', () => {
         endDate: null,
         id: 'f1288f6c-8503-4dc1-b114-75c408a14bd0',
         licenceHolder: 'Unregistered licence',
-        licenceName: 'Unregistered licence',
-        licenceRef: '01/123',
         monitoringStations: [{
           gaugingStationId: 'ac075651-4781-4e24-a684-b943b98607ca',
           label: 'MEVAGISSEY FIRE STATION'
         }],
-        notification: null,
-        pageTitle: 'Licence 01/123',
         purposes: null,
         region: 'Narnia',
-        registeredTo: null,
-        roles: null,
         sourceOfSupply: 'SURFACE WATER SOURCE OF SUPPLY',
-        startDate: '1 April 2019',
-        warning: null
+        startDate: '1 April 2019'
       })
     })
   })
 
-  describe("the 'abstractionConditionDetails' property", () => {
+  describe('the \'abstractionConditionDetails\' property', () => {
     describe('when there are multiple abstraction conditions', () => {
       beforeEach(() => {
         licenceAbstractionConditions.conditions = ['Derogation clause', 'General conditions']
@@ -68,7 +59,7 @@ describe('View Licence presenter', () => {
 
       describe('and they have different display titles', () => {
         it('returns the details with plural text and a populated conditions array', () => {
-          const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+          const result = ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
           expect(result.abstractionConditionDetails).to.equal({
             conditions: ['Derogation clause', 'General conditions'],
@@ -84,7 +75,7 @@ describe('View Licence presenter', () => {
         })
 
         it('returns the details with plural text and a populated conditions array', () => {
-          const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+          const result = ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
           expect(result.abstractionConditionDetails).to.equal({
             conditions: ['Derogation clause'],
@@ -101,7 +92,7 @@ describe('View Licence presenter', () => {
       })
 
       it('returns the details with singular text and a populated conditions array', () => {
-        const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.abstractionConditionDetails).to.equal({
           conditions: ['Derogation clause'],
@@ -118,7 +109,7 @@ describe('View Licence presenter', () => {
       })
 
       it('returns the details with plural text and an empty conditions array', () => {
-        const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.abstractionConditionDetails).to.equal({
           conditions: [],
@@ -128,10 +119,10 @@ describe('View Licence presenter', () => {
     })
   })
 
-  describe("the 'endDate' property", () => {
+  describe('the \'endDate\' property', () => {
     describe('when the licence expired date is null', () => {
       it('returns NULL', () => {
-        const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.endDate).to.be.null()
       })
@@ -149,7 +140,7 @@ describe('View Licence presenter', () => {
       })
 
       it('returns NULL', () => {
-        const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.endDate).to.be.null()
       })
@@ -160,18 +151,18 @@ describe('View Licence presenter', () => {
         licence.expiredDate = new Date('2099-04-01')
       })
 
-      it("returns '1 April 2099'", () => {
-        const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+      it('returns \'1 April 2099\'', () => {
+        const result = ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.endDate).to.equal('1 April 2099')
       })
     })
   })
 
-  describe("the 'licenceHolder' property", () => {
+  describe('the \'licenceHolder\' property', () => {
     describe('when the licence holder is not set', () => {
-      it("returns 'Unregistered licence'", () => {
-        const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+      it('returns \'Unregistered licence\'', () => {
+        const result = ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.licenceHolder).to.equal('Unregistered licence')
       })
@@ -182,40 +173,18 @@ describe('View Licence presenter', () => {
         licence.licenceHolder = 'Barbara Liskov'
       })
 
-      it("returns 'Barbara Liskov'", () => {
-        const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+      it('returns \'Barbara Liskov\'', () => {
+        const result = ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.licenceHolder).to.equal('Barbara Liskov')
       })
     })
   })
 
-  describe("the 'licenceName' property", () => {
-    describe('when there is no licenceName property', () => {
-      it('returns Unregistered licence', () => {
-        const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions)
-
-        expect(result.licenceName).to.equal('Unregistered licence')
-      })
-    })
-
-    describe('when there is a licenceName property', () => {
-      beforeEach(() => {
-        licence.licenceName = 'example@example.com'
-      })
-
-      it('returns a string with the licence name values', () => {
-        const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions)
-
-        expect(result.licenceName).to.equal('example@example.com')
-      })
-    })
-  })
-
-  describe("the 'licenceVersionPurposes' property", () => {
+  describe('the \'licenceVersionPurposes\' property', () => {
     describe('when there are no licenceVersions', () => {
       it('returns null', () => {
-        const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.abstractionPeriods).to.equal(null)
         expect(result.abstractionPeriodsAndPurposesLinkText).to.equal(null)
@@ -228,7 +197,7 @@ describe('View Licence presenter', () => {
       })
 
       it('returns null', () => {
-        const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.abstractionPeriods).to.equal(null)
         expect(result.abstractionPeriodsAndPurposesLinkText).to.equal(null)
@@ -246,7 +215,7 @@ describe('View Licence presenter', () => {
       })
 
       it('returns null', () => {
-        const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.abstractionPeriods).to.equal(null)
       })
@@ -268,7 +237,7 @@ describe('View Licence presenter', () => {
       })
 
       it('returns an object with a caption and an array with one abstraction period', () => {
-        const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.abstractionPeriods).to.equal({
           caption: 'Period of abstraction',
@@ -295,7 +264,7 @@ describe('View Licence presenter', () => {
       })
 
       it('returns an object with a caption and an array with one abstraction period', () => {
-        const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.abstractionPeriods).to.equal({
           caption: 'Period of abstraction',
@@ -327,7 +296,7 @@ describe('View Licence presenter', () => {
       })
 
       it('returns an object with a caption and an array with two purposes', () => {
-        const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.abstractionPeriods).to.equal({
           caption: 'Periods of abstraction',
@@ -337,14 +306,14 @@ describe('View Licence presenter', () => {
     })
   })
 
-  describe("the 'monitoringStations' property", () => {
+  describe('the \'monitoringStations\' property', () => {
     describe('when the licenceGaugingStations property is not an array', () => {
       beforeEach(() => {
         licence.licenceGaugingStations = {}
       })
 
       it('will return an empty array of monitoring station details', async () => {
-        const result = await ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = await ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.monitoringStations).to.equal([])
       })
@@ -356,19 +325,19 @@ describe('View Licence presenter', () => {
       })
 
       it('will return an empty array of monitoring station details', async () => {
-        const result = await ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = await ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.monitoringStations).to.equal([])
       })
     })
 
-    describe("when the licence has a null 'licenceGaugingStations' property", () => {
+    describe('when the licence has a null \'licenceGaugingStations\' property', () => {
       beforeEach(() => {
         licence.licenceGaugingStations = null
       })
 
       it('will return an empty array of monitoring station details', async () => {
-        const result = await ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = await ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.monitoringStations).to.equal([])
       })
@@ -376,7 +345,7 @@ describe('View Licence presenter', () => {
 
     describe('when the licence has a gauging station', () => {
       it('will return an array populated with monitoring station details', async () => {
-        const result = await ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = await ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.monitoringStations).to.equal([{
           gaugingStationId: 'ac075651-4781-4e24-a684-b943b98607ca',
@@ -397,7 +366,7 @@ describe('View Licence presenter', () => {
       })
 
       it('will return an array populated with multiple monitoring station details', async () => {
-        const result = await ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = await ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.monitoringStations).to.equal([{
           gaugingStationId: 'ac075651-4781-4e24-a684-b943b98607ca',
@@ -410,10 +379,10 @@ describe('View Licence presenter', () => {
     })
   })
 
-  describe("the 'purposes' property", () => {
+  describe('the \'purposes\' property', () => {
     describe('when there are no licenceVersions', () => {
       it('returns null', () => {
-        const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.purposes).to.equal(null)
       })
@@ -425,7 +394,7 @@ describe('View Licence presenter', () => {
       })
 
       it('returns null', () => {
-        const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.purposes).to.equal(null)
       })
@@ -439,7 +408,7 @@ describe('View Licence presenter', () => {
       })
 
       it('returns null', () => {
-        const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.purposes).to.equal(null)
       })
@@ -455,7 +424,7 @@ describe('View Licence presenter', () => {
       })
 
       it('returns an object with a caption and an array with one purpose', () => {
-        const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.purposes).to.equal({
           caption: 'Purpose',
@@ -476,7 +445,7 @@ describe('View Licence presenter', () => {
       })
 
       it('returns an object with a caption and an array with one entry', () => {
-        const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.purposes).to.equal({
           caption: 'Purpose',
@@ -497,7 +466,7 @@ describe('View Licence presenter', () => {
       })
 
       it('returns an object with a caption and an array with two entries', () => {
-        const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.purposes).to.equal({
           caption: 'Purposes',
@@ -507,32 +476,10 @@ describe('View Licence presenter', () => {
     })
   })
 
-  describe("the 'registeredTo' property", () => {
-    describe('when there is no registeredTo property', () => {
-      it('returns null', () => {
-        const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions)
-
-        expect(result.registeredTo).to.equal(null)
-      })
-    })
-
-    describe('when there is a registeredTo property', () => {
-      beforeEach(() => {
-        licence.registeredTo = 'Company'
-      })
-
-      it('returns a string with the registered to name', () => {
-        const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions)
-
-        expect(result.registeredTo).to.equal('Company')
-      })
-    })
-  })
-
-  describe("the 'purposes' property", () => {
+  describe('the \'purposes\' property', () => {
     describe('and it has a source of supply', () => {
       it('will return the source of supply for use in the licence summary page', async () => {
-        const result = await ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = await ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.sourceOfSupply).to.equal('SURFACE WATER SOURCE OF SUPPLY')
       })
@@ -550,7 +497,7 @@ describe('View Licence presenter', () => {
       })
 
       it('will return null for the source of supply', async () => {
-        const result = await ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = await ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.sourceOfSupply).to.equal(null)
       })
@@ -566,7 +513,7 @@ describe('View Licence presenter', () => {
       })
 
       it('will return null for the source of supply and abstraction point information', async () => {
-        const result = await ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = await ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.abstractionPoints).to.equal(null)
         expect(result.abstractionPointsCaption).to.equal('Point of abstraction')
@@ -585,7 +532,7 @@ describe('View Licence presenter', () => {
       })
 
       it('will return null for the source of supply and abstraction point information', async () => {
-        const result = await ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = await ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.abstractionPoints).to.equal(null)
         expect(result.abstractionPointsCaption).to.equal(null)
@@ -603,7 +550,7 @@ describe('View Licence presenter', () => {
       })
 
       it('will return null for the source of supply and abstraction point information', async () => {
-        const result = await ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = await ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.abstractionPoints).to.equal(null)
         expect(result.abstractionPointsCaption).to.equal(null)
@@ -621,7 +568,7 @@ describe('View Licence presenter', () => {
       })
 
       it('will return null for the source of supply and abstraction point information', async () => {
-        const result = await ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = await ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.abstractionPoints).to.equal(null)
         expect(result.abstractionPointsCaption).to.equal(null)
@@ -637,7 +584,7 @@ describe('View Licence presenter', () => {
       })
 
       it('will return null for the source of supply and abstraction point information', async () => {
-        const result = await ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = await ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.abstractionPoints).to.equal(null)
         expect(result.abstractionPointsCaption).to.equal(null)
@@ -653,7 +600,7 @@ describe('View Licence presenter', () => {
       })
 
       it('will return null for the source of supply and abstraction point information', async () => {
-        const result = await ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = await ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.abstractionPoints).to.equal(null)
         expect(result.abstractionPointsCaption).to.equal(null)
@@ -682,7 +629,7 @@ describe('View Licence presenter', () => {
       })
 
       it('will return the correct information for the abstraction point', async () => {
-        const result = await ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = await ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.abstractionPoints).to.equal([
           'Between National Grid References TL 23198 88603 and TM 23197 88602'
@@ -711,7 +658,7 @@ describe('View Licence presenter', () => {
       })
 
       it('will return the information for the abstraction point', async () => {
-        const result = await ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = await ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.abstractionPoints).to.equal([
           'Within the area formed by the straight lines running between National Grid References TL 23198 88603 TM 23197 88602 TN 23196 88601 and TO 23195 88600'
@@ -734,7 +681,7 @@ describe('View Licence presenter', () => {
       })
 
       it('will return the information for the abstraction point', async () => {
-        const result = await ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = await ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.abstractionPoints).to.equal([
           'Between National Grid References TL 23198 88603 and TM 23197 88602'
@@ -757,7 +704,7 @@ describe('View Licence presenter', () => {
       })
 
       it('will return the information for the abstraction point', async () => {
-        const result = await ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = await ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.abstractionPoints).to.equal(['At National Grid Reference TL 23198 88603'])
         expect(result.abstractionPointsCaption).to.equal('Point of abstraction')
@@ -775,7 +722,7 @@ describe('View Licence presenter', () => {
       })
 
       it('will return the information for the abstraction point', async () => {
-        const result = await ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = await ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.abstractionPoints).to.equal([
           'At National Grid Reference TL 23198 88603'
@@ -796,7 +743,7 @@ describe('View Licence presenter', () => {
       })
 
       it('will return the information for the abstraction point', async () => {
-        const result = await ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = await ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.abstractionPoints).to.equal([
           'At National Grid Reference TL 23198 88603 (Local)'
@@ -823,7 +770,7 @@ describe('View Licence presenter', () => {
       })
 
       it('will return the information for the abstraction point', async () => {
-        const result = await ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = await ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.abstractionPoints).to.equal([
           'At National Grid Reference TL 23198 88603',
@@ -851,7 +798,7 @@ describe('View Licence presenter', () => {
       })
 
       it('will only display one of the abstraction point', async () => {
-        const result = await ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = await ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.abstractionPoints).to.equal([
           'At National Grid Reference TL 23198 88603'
@@ -870,7 +817,7 @@ describe('View Licence presenter', () => {
       })
 
       it('will display the formatted strings with the rates per period and the correct caption', async () => {
-        const result = await ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = await ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.abstractionQuantities).to.equal([
           '265.00 cubic metres per year',
@@ -887,7 +834,7 @@ describe('View Licence presenter', () => {
       })
 
       it('will display the formatted string with the rate per period and the correct caption', async () => {
-        const result = await ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = await ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.abstractionQuantities).to.equal([
           '265.00 cubic metres per year'
@@ -904,7 +851,7 @@ describe('View Licence presenter', () => {
       })
 
       it('will display the formatted string with the rate per period and the correct caption', async () => {
-        const result = await ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = await ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.abstractionQuantities).to.equal([
           '60.00 cubic metres per hour',
@@ -922,7 +869,7 @@ describe('View Licence presenter', () => {
       })
 
       it('will display the formatted string with the rate per period and the correct caption', async () => {
-        const result = await ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = await ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.abstractionQuantities).to.equal([
           '6.00 litres per second'
@@ -966,7 +913,7 @@ describe('View Licence presenter', () => {
       })
 
       it('will display the formatted string with the rate per period and the correct caption', async () => {
-        const result = await ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = await ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.abstractionQuantities).to.equal([
           '265.00 cubic metres per year',
@@ -1013,147 +960,9 @@ describe('View Licence presenter', () => {
       })
 
       it('will display the formatted string with the rate per period and the correct caption', async () => {
-        const result = await ViewLicencePresenter.go(licence, licenceAbstractionConditions)
+        const result = await ViewLicenceSummaryPresenter.go(licence, licenceAbstractionConditions)
 
         expect(result.abstractionQuantities).to.equal(null)
-      })
-    })
-  })
-
-  describe("the 'warning' property", () => {
-    describe('when the licence does not have an end date (expired, lapsed or revoked)', () => {
-      it('returns NULL', () => {
-        const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions)
-
-        expect(result.warning).to.be.null()
-      })
-    })
-
-    describe('when the licence does have an end date but it is in the future (expired, lapsed or revoked)', () => {
-      beforeEach(() => {
-        licence.expiredDate = new Date('2099-04-01')
-      })
-
-      it('returns NULL', () => {
-        const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions)
-
-        expect(result.warning).to.be.null()
-      })
-    })
-
-    describe('when the licence ends today or in the past (2019-04-01) because it is expired', () => {
-      beforeEach(() => {
-        licence.ends = { date: new Date('2019-04-01'), reason: 'expired' }
-      })
-
-      it("returns 'This licence expired on 1 April 2019'", () => {
-        const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions)
-
-        expect(result.warning).to.equal('This licence expired on 1 April 2019')
-      })
-    })
-
-    describe('when the licence ends today or in the past (2019-04-01) because it is lapsed', () => {
-      beforeEach(() => {
-        licence.ends = { date: new Date('2019-04-01'), reason: 'lapsed' }
-      })
-
-      it("returns 'This licence lapsed on 1 April 2019'", () => {
-        const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions)
-
-        expect(result.warning).to.equal('This licence lapsed on 1 April 2019')
-      })
-    })
-
-    describe('when the licence was ends today or in the past (2019-04-01) because it is revoked', () => {
-      beforeEach(() => {
-        licence.ends = { date: new Date('2019-04-01'), reason: 'revoked' }
-      })
-
-      it("returns 'This licence was revoked on 1 April 2019'", () => {
-        const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions)
-
-        expect(result.warning).to.equal('This licence was revoked on 1 April 2019')
-      })
-    })
-  })
-  describe("the 'notification' property", () => {
-    describe('when the licence will not be in the next supplementary bill run', () => {
-      it('returns NULL', () => {
-        const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions)
-
-        expect(result.notification).to.be.null()
-      })
-    })
-
-    describe('when the licence will be in the next supplementary bill run (PRESROC)', () => {
-      beforeEach(() => {
-        licence.includeInPresrocBilling = 'yes'
-      })
-
-      it('returns the notification for PRESROC', () => {
-        const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions)
-
-        expect(result.notification).to.equal('This license has been marked for the next supplementary bill run for the old charge scheme.')
-      })
-    })
-
-    describe('when the licence will be in the next supplementary bill run (SROC)', () => {
-      beforeEach(() => {
-        licence.includeInSrocBilling = true
-      })
-
-      it('returns the notification for SROC', () => {
-        const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions)
-
-        expect(result.notification).to.equal('This license has been marked for the next supplementary bill run.')
-      })
-    })
-
-    describe('when the licence will be in the next supplementary bill run (SROC & PRESROC)', () => {
-      beforeEach(() => {
-        licence.includeInSrocBilling = true
-        licence.includeInPresrocBilling = 'yes'
-      })
-
-      it('returns the notification for SROC & PRESROC)', () => {
-        const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions)
-
-        expect(result.notification).to.equal('This license has been marked for the next supplementary bill runs for the current and old charge schemes.')
-      })
-    })
-  })
-
-  describe("the 'roles' property", () => {
-    describe('when the authenticated user has roles', () => {
-      beforeEach(() => {
-        auth = {
-          credentials: {
-            roles: [{
-              role: 'role 1'
-            },
-            {
-              role: 'role 2'
-            }]
-          }
-        }
-      })
-
-      it('returns the roles in a flat array', () => {
-        const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions, auth)
-
-        expect(result.roles).to.equal(['role 1', 'role 2'])
-      })
-    })
-    describe('when the authenticated user has NO roles', () => {
-      beforeEach(() => {
-        auth = undefined
-      })
-
-      it('returns null for the roles', () => {
-        const result = ViewLicencePresenter.go(licence, licenceAbstractionConditions, auth)
-
-        expect(result.roles).to.be.null()
       })
     })
   })
